requirements.txt
django>=4.2
pandas
requests
python-decouple
